+++ react/SummaryCoupon.tsx
@@ -1,3 +1,16 @@
-import SummaryCoupon from './components/SummaryCoupon/SummaryCoupon'
+import React, { FunctionComponent } from 'react'
+import { ExtensionPoint } from 'vtex.render-runtime'
+
+import { OrderCouponProvider } from 'vtex.order-coupon/OrderCoupon'
+
+const SummaryCoupon: FunctionComponent<SummaryCouponProps> = () => {
+  return (
+    <OrderCouponProvider>
+      <ExtensionPoint id="coupon" />
+    </OrderCouponProvider>
+  )
+}
+
+interface SummaryCouponProps {}
 
 export default SummaryCoupon
