+++ react/components/SummaryItem.tsx
@@ -54,10 +54,10 @@ const SummaryItem: FunctionComponent<Props & InjectedIntlProps> = ({
     </div>
     <div
       className={`flex-auto tr ${large ? 'fw6 fw5-l' : ''} ${
-        value ? '' : 'c-success'
+        value === 0 ? 'c-success' : ''
       }`}
     >
-      <FormattedPrice value={value / 100}></FormattedPrice>
+      <FormattedPrice value={value ? value / 100 : value}></FormattedPrice>
     </div>
   </div>
 )
